+++ source/extensions/omni.isaac.extension_templates/docs/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## [1.7.0] - 2024-03-14
+### Fixed
+- Fixed code for handling Articulation selection in Configuration Tooling Template.
+    The issue arose from a change in behavior in Core get_prim_object_type()
+
 ## [1.6.0] - 2024-02-28
 ### Fixed
 - Fixed code for populating drop-down menu with Articulations in Configuration Tooling Template.