+++ source/extensions/omni.isaac.gain_tuner/docs/CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
-## [1.0.0] - 2024-03-08
-
+## [1.0.1] - 2024-03-14
+### Fixed
+- Fixed logic around selecting Articulation on STOP/PLAY given new behavior in Core
+  get_prim_object_type() function.
 ### Added
+- Added helper text telling the user to click the play button to get started.
+- Added more text clarifying Gains Test settings.
 
+## [1.0.0] - 2024-03-08
+### Added
 - Initial version of Gain Tuner Extension