+++ source/extensions/omni.isaac.lula_test_widget/docs/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## [0.2.2] - 2024-03-15
+### Fixed
+- Fixed logic around selecting Articulation on STOP/PLAY given new behavior in Core
+  get_prim_object_type() function.
+
 ## [0.2.1] - 2023-11-13
 ### Fixed
 - Updated documentation link